created Processing class, calculate average

+++ com.viatra.cps.benchmark.reports.processing/src/com/viatra/cps/benchmark/reports/processing/models/Metric.java
@@ -2,9 +2,9 @@ package com.viatra.cps.benchmark.reports.processing.models;
 
 public class Metric {
 	private String name;
-	private Long value;
+	private Double value;
 
-	public Metric(String name, Long value) {
+	public Metric(String name, Double value) {
 		this.name = name;
 		this.value = value;
 	}
@@ -13,10 +13,10 @@ public class Metric {
 		return name;
 	}
 
-	public Long getValue() {
+	public Double getValue() {
 		return value;
 	}
-	
+
 	public void print() {
 		System.out.println("\t\t\tMetric: \n\t\t\t{");
 		System.out.println("\t\t\t\tMetricName: " + name);
